requirements.txt
cssselect==0.9.1
dj-database-url==0.4.1
Django==1.11
django-filter==0.14.0
djangorestframework==3.4.6
flup==1.0.2
gunicorn==19.6.0
lxml==3.4.2
Markdown==2.6.6
pymongo==2.8
pyquery==1.2.9
python-dateutil==2.4.1
pytz==2016.7
six==1.9.0
whitenoise==3.2.1
django-treebeard==4.1.0
markdown2==2.3.1
bleach==1.5.0
